bot/pyproject.toml
[tool.poetry]
name = "hoardbooru-bot"
version = "0.1.0"
description = "Telegram bot for interfacing with hoardbooru"
authors = ["Deer-Spangle <deer@spangle.org.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.38.1"
prometheus-client = "^0.21.1"
pyszuru = "^0.3.1"
aiohttp = "^3.11.10"
aiosqlite = "^0.20.0"
aiofiles = "^24.1.0"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
